refactor: extract settings routes, test real POST handler
Extract GET/POST/DELETE /settings routes from server.ts into
routes/settingsRoutes.ts. Side-effect dependencies (meshtasticManager,
inactiveNodeNotificationService, etc.) are injected via callbacks.

The persistence test now mounts the real settingsRoutes router instead
of a mock reimplementation, so it tests the actual filtering, validation,
and save codepath. Test values are generated via validTestValue() to
satisfy the real handler's validators.

Also replaces hardcoded SETTINGS_TAB_SENDS and SETTINGS_CONTEXT_LOADS
arrays with regex extraction from the frontend source files at test time,
catching frontend↔server desyncs automatically.
